requirements.txt
semantic-kernel 
pyautogen